init-lcow: Update build to latest OpenGCS
The opengcs build has changed considerably. There is now a top
level Makefile and three binaries we need to copy. The other
binaries are symlinks.

Also use the 'init' script from OpenGCS repository, which is
a copy of the 'init' script we had in this repository (with
some minor modifications).

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>